requirements.txt
mysql-connector-python
python-dotenv
pytest
flake8
black